UI performance + other upgrades
From now on, the web UI does not accept raw SST files, but PSST files
preprocessed by the gosst utility. Dynamic calculations (histograms and
FFT graphs) are done in Python, but they are quite fast, due to being
done with Numpy.

The travel graph now displays a subsampled (every 100th data point) data
set, because add_trace is too slow. Underlying calculations are done on
the whole dataset of course.

We are now displaying travel and velocity histograms besides the travel
and FFT graphs. Some statistic calculations that should help finding
correct tuning settings are also performed and the results are placed on
the corresponding graphs.