[ui] Filter velocity data by minimum stroke length
As part of the balance calculations, we are classifying telemetry data
into compression and rebound strokes. This can also be used to filter
data further (in addition to the Savitzky-Golay filter, which is used
to compute the first derivative of the travel data), i.e. we can ignore
strokes that are small enough to be considered noise. For now, this
threshold is 5 mm.