[ui] Exclude right edge from interval masks
Interval start / end values are used directly to get subarrays, so
it is necessary to exclude the end value from masks to be consistent
with how the array[start:end] notation works.

The old behaviour introduced a bug by causing rebound velocity values
to appear in compression strokes and vice versa.