[gosst] Switch to stroke-based calculations
Instead of masking out jumps and idlings on the Python side, we filter
and categorize individual strokes as either compressions, rebounds,
airtimes or idlings during preprocessing.

Balance data and histograms are also calculated here individually for
each compression and rebound stroke, so that we can still calculate
plot data dynamically for a selected interval, but with much less
calculation on the Python side. This should also significantly decrease
code complexity of the dashboard.