[api,ui] Fix database migrations
Flask-Migrate was introduced after version v0.2.0-alpha, and the first
migration was created when the token_blocklist table was added. This was
a mistake, since a "flask db upgrade" creates only that table when run
on an empty database. So the original migrations directory was scrapped
and recreated with an initial migration with all the tables in version
v0.2.0-alpha, and another one with token_blocklist. With this change, we
can now call flask_migrate.upgrade() instead of db.create_all() when the
database does not exists yet.

Existing v0.2.0-alpha databases do not have a revision (they do not have
an alembic_version table), so in order for them to work nicely, we
needed to stamp them to the initial migration revision.